+++ packages/fa_ui/lib/src/chat/fa_chat_screen.dart
@@ -74,6 +74,7 @@ class FaChatScreen extends StatefulWidget {
     this.wallpaperBuilder,
     this.avatarBuilder,
     this.onAuthRecovery,
+    this.onPermissionAction,
     this.audioControllerFactory,
     this.videoControllerFactory,
   });